Refactor partitioner approaches Clearer naming
