+++ superblockify/partitioning/approaches/__init__.py
@@ -1,0 +1,4 @@
+"""Approaches init, subpackage for the various approaches."""
+from .bearing import BearingPartitioner
+from .dummy import DummyPartitioner
+from .streettype import ResidentialPartitioner
