+++ superblockify/partitioning/approaches/dummy.py
@@ -3,7 +3,7 @@ from networkx import weakly_connected_components
 from numpy import mean, min as npmin, max as npmax
 from typing_extensions import deprecated
 
-from .partitioner import BasePartitioner
+from ..base import BasePartitioner
 
 
 @deprecated("DummyPartitioner does not necessarily produce a valid partitioning.")
